dbaas: pg connection pool (#510)
# Description
<!--
* Prefix: the title with the component name being changed. Add a short
and self describing sentence to ease the review
* Please add a few lines providing context and describing the change
* Please self comment changes whenever applicable to help with the
review process
* Please keep the checklist as part of the PR. Tick what applies to this
change.
-->

## Checklist
(For exoscale contributors)

* [x] Changelog updated (under *Unreleased* block)
* [x] [Acceptance tests
OK](https://github.com/exoscale/terraform-provider-exoscale/actions/runs/23840611720)
* [x] For a new resource, datasource or new attributes: acceptance test
added/updated

## Testing

Tested local build with pre-prod.

+++ pkg/resources/database/resource_db_pg.go
@@ -256,6 +256,26 @@ func (data PGDatabaseResourceModel) CreateResource(ctx context.Context, client *
 
 // DeleteResource deletes the resource
 func (data PGDatabaseResourceModel) DeleteResource(ctx context.Context, client *v3.Client, diagnostics *diag.Diagnostics) {
+	blockingPools, err := getPGConnectionPoolsUsingDatabase(ctx, client, data.Service.ValueString(), data.DatabaseName.ValueString())
+	if err != nil {
+		diagnostics.AddError(
+			"Client Error",
+			fmt.Sprintf("unable to check connection pool references for service database %q: %s", data.DatabaseName.ValueString(), err.Error()),
+		)
+		return
+	}
+
+	if len(blockingPools) > 0 {
+		diagnostics.AddError(
+			"Client Error",
+			fmt.Sprintf(
+				"service database %q is referenced by connection pool(s) %s; delete or update the connection pool(s) first",
+				data.DatabaseName.ValueString(),
+				formatPGConnectionPoolReferences(blockingPools),
+			),
+		)
+		return
+	}
 
 	op, err := client.DeleteDBAASPGDatabase(ctx, data.Service.ValueString(), data.DatabaseName.ValueString())
 	if err != nil {
